robot/requirements.txt
PySide6
PyQt6
qasync
aiohttp
opencv-contrib-python-headless
cv_bridge
pyserial
numpy<2
httpx
python-dotenv
fastapi
pydantic
st3215
pytest-asyncio>=0.23.0